requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
black==23.9.1
blinker==1.6.2
boto3==1.34.2
botocore==1.34.2
bs4==0.0.1
CD==1.1.0
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
dataclasses-json==0.6.1
dnspython==2.4.2
email-validator==2.1.0.post1
emoji==2.8.0
faiss-cpu==1.7.4
fastapi==0.103.2
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-mongoengine==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fpdf==1.7.2
frozenlist==1.4.0
fsspec==2023.12.2
gdown==4.7.1
google-search-results==2.4.2
html5lib==1.1
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.305
langdetect==1.0.9
langsmith==0.0.41
lxml==4.9.3
make-response==1
MarkupSafe==2.1.3
marshmallow==3.20.1
mongoengine==0.27.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.2
openai==0.28.1
opencv-python==4.8.1.78
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.10.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymongo==4.6.1
pypdf==3.16.2
PyPDF2==3.0.1
PySocks==1.7.1
python-amazon-paapi==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
s3transfer==0.9.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
unstructured==0.10.18
urllib3==2.0.5
webencodings==0.5.1
Werkzeug==3.0.0
WTForms==3.1.1
yarl==1.9.2
